Implement ELP crash basis with theory-based variable weights
Replace the old 0-5 integer weight scheme with a new scheme that reflects
the theoretical structure of the extended LP at each partition refinement:
- INFINITY: basic non-degenerate variables and all their children (highest priority)
- 5: r-variables (residual/linking columns) and their row slack children
- 4: degenerate basic variables and all their children
- 0: nonbasic variables and their children

Weight array resized from colCnt+rowCnt (aggregated LP) to
colCnt+numResiduals+rowCnt+numResiduals to cover the full ELP.
Row slack offset shifted by numResiduals accordingly.
Two new functions: buildResidualCrashBasisWeights() and
buildResidualRowCrashBasisWeights().

colweights type changed from std::vector<int> to std::vector<double>
throughout (OCAggregate, Highs, ICrashX, lp_solver) so INFINITY flows
correctly into IPX GuessBasis.

Switch crash/crossover from aggregated LP (alp_) to extended LP (ealp_):
- buildOCALP() now also sets ealp_ = aggregator_.getLp()
- buildSolution() resized to colCnt+numResiduals / rowCnt+numResiduals
  (r-variable entries initialize to 0.0 = at lower bound)
- All 8 passModel and crossover/primalCrossover call sites updated to ealp_

Tested on all small instances; all return Optimal with no errors.

+++ HiGHS-1-2-1/src/Highs.h
@@ -877,8 +877,8 @@ class Highs {
   HighsStatus crossover(HighsSolution& solution);
   HighsStatus crossover(HighsSolution& solution, HighsLp& lp);
   HighsStatus primalCrossover(HighsSolution& solution, HighsLp& lp);
-  HighsStatus crossover(HighsSolution& solution, HighsLp& lp, std::vector<int>& colweights);
-  HighsStatus primalCrossover(HighsSolution& solution, HighsLp& lp, std::vector<int>& colweights);
+  HighsStatus crossover(HighsSolution& solution, HighsLp& lp, std::vector<double>& colweights);
+  HighsStatus primalCrossover(HighsSolution& solution, HighsLp& lp, std::vector<double>& colweights);
 
   /**
    * @brief Open a named log file
@@ -913,7 +913,7 @@ class Highs {
   HighsLp ealp_;
   HighsLp pealp_;
   HighsBasis alpBasis_;
-  std::vector<int> alpColweights_;
+  std::vector<double> alpColweights_;
   HighsBasis crashBasis_;
   HighsSolution crashSolution_;
   HighsSolution alpSolution_;
